--- a/resources/views/components/includes/tab_navs/settings-privacy-navs.blade.php
+++ b/resources/views/components/includes/tab_navs/settings-privacy-navs.blade.php
@@ -8,11 +8,11 @@
     [
         'title' => 'Security',
         'icon' => 'shield-plus',
-        'route' => 'settings',
+        'route' => 'profile.settings',
     ],
     [
         'title' => 'Activity Log',
         'icon' => 'list',
-        'route' => 'activity-logs',
+        'route' => 'profile.logs',
     ],
 ]" />