### 🛠 Changes

- Cache employee information to reduce performance overhead on database operations. This drastically reduce the needs to query employee information everytime `/profile` or `/employee` information is accessed.

without cache

![image](https://github.com/user-attachments/assets/6bcd24bf-9abc-4ed0-a0dc-d9f0e08aa1fe)


with cache

![image](https://github.com/user-attachments/assets/8113f0eb-170b-4bf2-b18e-1e4b04f332a1)

- As you can see, the elapsed time is drastically reduced for like 4x from `466ms` to `82.02ms` and the select queries is cut more than half from `26` to `12`. Latency went from `1.46s` to `454ms`

### ✔️ Checklist

- [x] Is `staging` the base branch of this PR?
- [x] Do your changes not reveal sensitive information, such as secrets, API keys, etc?
- [x] Are there no erroneous console logs, debuggers, or leftover code in your changes?